pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.45.2",
    "ccxt>=4.4.57",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "numpy>=2.2.2",
    "oauthlib>=3.2.2",
    "openai>=1.61.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "slack-sdk>=3.34.0",
    "trafilatura>=2.0.0",
    "twilio>=9.4.4",
    "urllib3>=2.3.0",
]
